Updated patterns to support base64 data urls
I had a use case where I'm using the slider in a preview before the user uploads files to the server, so I only have base64 data urls.

The loader would just hang because it wouldn't know what to do with them when I tried to navigate to the second image in the list.

I simply updated the pattern matching to understand the base64's prefix in addition to the file extensions it already supported.